+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Adds repeating awards, use % in front of the play count to indicate repeating award.
+- Adds user tracking for awards.
+
 ## [v5.28.2](https://github.com/lexicalunit/spellbot/releases/tag/v5.28.2) - 2021-04-20
 
 ### Fixed